+++ modules/app-desktop/src/main/kotlin/org/archivekeep/app/desktop/ui/components/operations/IndexUpdatePreparationProgress.kt
@@ -4,8 +4,8 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import org.archivekeep.app.desktop.ui.designsystem.progress.ProgressRow
 import org.archivekeep.app.desktop.ui.designsystem.progress.ProgressRowList
-import org.archivekeep.app.desktop.ui.utils.filesAutoPlural
 import org.archivekeep.files.operations.indexupdate.AddOperation
+import org.archivekeep.utils.filesAutoPlural
 
 @Composable
 fun IndexUpdatePreparationProgress(progress: AddOperation.PreparationProgress) {