+++ modules/app-desktop/src/main/kotlin/org/archivekeep/app/desktop/ui/dialogs/repository/operations/addpush/AddAndPushRepoDialog.kt
@@ -42,13 +42,13 @@ import org.archivekeep.app.desktop.ui.designsystem.progress.ProgressRowList
 import org.archivekeep.app.desktop.ui.dialogs.repository.AbstractRepositoryDialog
 import org.archivekeep.app.desktop.ui.dialogs.repository.operations.addpush.AddAndPushRepoDialogViewModel.VMState
 import org.archivekeep.app.desktop.ui.utils.contextualStorageReference
-import org.archivekeep.app.desktop.ui.utils.filesAutoPlural
 import org.archivekeep.app.desktop.utils.asMutableState
 import org.archivekeep.app.desktop.utils.collectAsLoadable
 import org.archivekeep.files.operations.indexupdate.AddOperation
 import org.archivekeep.files.operations.indexupdate.IndexUpdateAddProgress
 import org.archivekeep.files.operations.indexupdate.IndexUpdateMoveProgress
 import org.archivekeep.utils.collections.ifNotEmpty
+import org.archivekeep.utils.filesAutoPlural
 import org.archivekeep.utils.loading.Loadable
 import org.archivekeep.utils.loading.mapIfLoadedOrDefault