+++ modules/app-desktop/src/main/kotlin/org/archivekeep/app/desktop/ui/utils/texts.kt
@@ -10,10 +10,3 @@ fun contextualStorageReference(
 } else {
     storageRepository.storage.displayName
 }
-
-fun filesAutoPlural(items: Collection<*>): String =
-    if (items.size == 1) {
-        "file"
-    } else {
-        "files"
-    }