+++ modules/files/src/main/kotlin/org/archivekeep/utils/texts.kt
@@ -1,0 +1,8 @@
+package org.archivekeep.utils
+
+fun filesAutoPlural(items: Collection<*>): String =
+    if (items.size == 1) {
+        "file"
+    } else {
+        "files"
+    }